+++ examples/standalone-hibernate-example/src/main/java/br/com/dillmann/dynamicquery/examples/hibernate/productgroup/ProductGroup.java
@@ -1,0 +1,30 @@
+package br.com.dillmann.dynamicquery.examples.hibernate.productgroup;
+
+import jakarta.persistence.Entity;
+import jakarta.persistence.Id;
+import jakarta.persistence.Table;
+
+@Entity
+@Table(name = "product_group")
+public class ProductGroup {
+
+    @Id
+    private Integer id;
+    private String description;
+
+    public Integer getId() {
+        return id;
+    }
+
+    public void setId(Integer id) {
+        this.id = id;
+    }
+
+    public String getDescription() {
+        return description;
+    }
+
+    public void setDescription(String description) {
+        this.description = description;
+    }
+}