+++ examples/standalone-hibernate-example/src/main/java/br/com/dillmann/dynamicquery/examples/hibernate/productgroup/ProductGroupHttpHandler.java
@@ -1,0 +1,10 @@
+package br.com.dillmann.dynamicquery.examples.hibernate.productgroup;
+
+import br.com.dillmann.dynamicquery.examples.hibernate.AbstractHttpHandler;
+import jakarta.persistence.EntityManager;
+
+public class ProductGroupHttpHandler extends AbstractHttpHandler<ProductGroup> {
+    public ProductGroupHttpHandler(final EntityManager entityManager) {
+        super(entityManager, ProductGroup.class);
+    }
+}